pyproject.toml
[tool.poetry]
name = "shopify-webhook-parser-lib"
version = "0.1.0"
description = ""
authors = ["Zin Yosrim <zinyosrim@gmail.com>"]
readme = "README.md"
packages = [{include = "shopifywebhookparser"}]

[tool.poetry.dependencies]
python = "^3.9"
azure-functions = "^1.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
